Create NodeIdenticon component
