+++ views/Settings/Nodes.tsx
@@ -1,10 +1,11 @@
 import * as React from 'react';
 import { FlatList, View } from 'react-native';
-import { Avatar, Header, Icon, ListItem } from 'react-native-elements';
+import { Header, Icon, ListItem } from 'react-native-elements';
 import Identicon from 'identicon.js';
 import { inject, observer } from 'mobx-react';
 
 import Button from './../../components/Button';
+import NodeIdenticon, { NodeTitle } from './../../components/NodeIdenticon';
 import SettingsStore from './../../stores/SettingsStore';
 import PrivacyUtils from './../../utils/PrivacyUtils';
 import { localeString } from './../../utils/LocaleUtils';
@@ -73,15 +74,6 @@ export default class Nodes extends React.Component<NodesProps, NodesState> {
         const { setSettings, settings }: any = SettingsStore;
         const { selectedNode } = settings;
 
-        const Node = (balanceImage: string) => (
-            <Avatar
-                source={{
-                    uri: balanceImage
-                }}
-                rounded
-            />
-        );
-
         const BackButton = () => (
             <Icon
                 name="arrow-back"
@@ -110,113 +102,79 @@ export default class Nodes extends React.Component<NodesProps, NodesState> {
                     {!!nodes && nodes.length > 0 && (
                         <FlatList
                             data={nodes}
-                            renderItem={({ item, index }) => {
-                                const displayName = item.nickname
-                                    ? item.nickname
-                                    : item.implementation === 'lndhub'
-                                    ? item.lndhubUrl
-                                          .replace('https://', '')
-                                          .replace('http://', '')
-                                    : item.url
-                                    ? item.url
-                                          .replace('https://', '')
-                                          .replace('http://', '')
-                                    : item.port
-                                    ? `${item.host}:${item.port}`
-                                    : item.host || 'Unknown';
-
-                                const title = PrivacyUtils.sensitiveValue(
-                                    displayName,
-                                    8
-                                );
-                                const implementation =
-                                    PrivacyUtils.sensitiveValue(
-                                        item.implementation || 'lnd',
-                                        8
-                                    );
-
-                                const data = new Identicon(
-                                    hash.sha1(
-                                        item.implementation === 'lndhub'
-                                            ? `${title}-${item.username}`
-                                            : title
-                                    ),
-                                    255
-                                ).toString();
-
-                                return (
-                                    <ListItem
-                                        title={`${title}`}
-                                        leftElement={Node(
-                                            `data:image/png;base64,${data}`
-                                        )}
-                                        rightElement={
-                                            <Button
-                                                title=""
-                                                icon={{
-                                                    name: 'settings',
-                                                    size: 25,
-                                                    color: themeColor('text')
-                                                }}
-                                                buttonStyle={{
-                                                    backgroundColor:
-                                                        'transparent',
-                                                    marginRight: -10
-                                                }}
-                                                onPress={() =>
-                                                    navigation.navigate(
-                                                        'AddEditNode',
-                                                        {
-                                                            node: item,
-                                                            index: index,
-                                                            active:
-                                                                selectedNode ===
-                                                                index,
-                                                            saved: true
-                                                        }
-                                                    )
-                                                }
-                                                iconOnly
-                                                adaptiveWidth
-                                            />
-                                        }
-                                        subtitle={
-                                            selectedNode === index ||
-                                            (!selectedNode && index === 0)
-                                                ? `Active | ${implementation}`
-                                                : `${implementation}`
-                                        }
-                                        containerStyle={{
-                                            borderBottomWidth: 0,
-                                            backgroundColor:
-                                                themeColor('background')
-                                        }}
-                                        onPress={() => {
-                                            setSettings(
-                                                JSON.stringify({
-                                                    nodes,
-                                                    theme: settings.theme,
-                                                    selectedNode: index,
-                                                    fiat: settings.fiat,
-                                                    passphrase:
-                                                        settings.passphrase,
-                                                    privacy: settings.privacy
-                                                })
-                                            ).then(() => {
-                                                navigation.navigate('Wallet', {
-                                                    refresh: true
-                                                });
+                            renderItem={({ item, index }) => (
+                                <ListItem
+                                    title={`${NodeTitle(item)}`}
+                                    leftElement={
+                                        <NodeIdenticon
+                                            selectedNode={item}
+                                            width={35}
+                                        />
+                                    }
+                                    rightElement={
+                                        <Button
+                                            title=""
+                                            icon={{
+                                                name: 'settings',
+                                                size: 25,
+                                                color: themeColor('text')
+                                            }}
+                                            buttonStyle={{
+                                                backgroundColor: 'transparent',
+                                                marginRight: -10
+                                            }}
+                                            onPress={() =>
+                                                navigation.navigate(
+                                                    'AddEditNode',
+                                                    {
+                                                        node: item,
+                                                        index: index,
+                                                        active:
+                                                            selectedNode ===
+                                                            index,
+                                                        saved: true
+                                                    }
+                                                )
+                                            }
+                                            iconOnly
+                                            adaptiveWidth
+                                        />
+                                    }
+                                    subtitle={
+                                        selectedNode === index ||
+                                        (!selectedNode && index === 0)
+                                            ? `Active | ${implementation}`
+                                            : `${implementation}`
+                                    }
+                                    containerStyle={{
+                                        borderBottomWidth: 0,
+                                        backgroundColor:
+                                            themeColor('background')
+                                    }}
+                                    onPress={() => {
+                                        setSettings(
+                                            JSON.stringify({
+                                                nodes,
+                                                theme: settings.theme,
+                                                selectedNode: index,
+                                                fiat: settings.fiat,
+                                                passphrase: settings.passphrase,
+                                                privacy: settings.privacy
+                                            })
+                                        ).then(() => {
+                                            navigation.navigate('Wallet', {
+                                                refresh: true
                                             });
-                                        }}
-                                        titleStyle={{
-                                            color: themeColor('text')
-                                        }}
-                                        subtitleStyle={{
-                                            color: themeColor('secondaryText')
-                                        }}
-                                    />
-                                );
-                            }}
+                                        });
+                                    }}
+                                    titleStyle={{
+                                        color: themeColor('text')
+                                    }}
+                                    subtitleStyle={{
+                                        color: themeColor('secondaryText')
+                                    }}
+                                />
+                            )}
                             refreshing={loading}
                             keyExtractor={(item, index) =>
                                 `${item.host}-${index}`
