+++ views/Settings/Settings.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { StyleSheet, Text, View, TouchableOpacity } from 'react-native';
-import { Avatar, Header, Icon } from 'react-native-elements';
+import { Header, Icon } from 'react-native-elements';
 import { inject, observer } from 'mobx-react';
 import Identicon from 'identicon.js';
 
@@ -16,6 +16,7 @@ import LanguageIcon from '../../images/SVG/Globe.svg';
 import HelpIcon from '../../images/SVG/Help Icon.svg';
 import NodeOn from '../../images/SVG/Node On.svg';
 
+import NodeIdenticon, { NodeTitle } from './../../components/NodeIdenticon';
 import { themeColor } from './../../utils/ThemeUtils';
 import { localeString } from './../../utils/LocaleUtils';
 import PrivacyUtils from './../../utils/PrivacyUtils';
@@ -64,46 +65,6 @@ export default class Settings extends React.Component<SettingsProps, {}> {
             />
         );
 
-        const displayName =
-            selectedNode && selectedNode.nickname
-                ? selectedNode.nickname
-                : selectedNode && selectedNode.implementation === 'lndhub'
-                ? selectedNode.lndhubUrl
-                      .replace('https://', '')
-                      .replace('http://', '')
-                : selectedNode && selectedNode.url
-                ? selectedNode.url
-                      .replace('https://', '')
-                      .replace('http://', '')
-                : selectedNode && selectedNode.port
-                ? `${selectedNode.host}:${selectedNode.port}`
-                : (selectedNode && selectedNode.host) || 'Unknown';
-
-        const title = PrivacyUtils.sensitiveValue(displayName, 8);
-        const implementation = PrivacyUtils.sensitiveValue(
-            (selectedNode && selectedNode.implementation) || 'lnd',
-            8
-        );
-
-        const data = new Identicon(
-            hash.sha1(
-                selectedNode && selectedNode.implementation === 'lndhub'
-                    ? `${title}-${selectedNode.username}`
-                    : title
-            ),
-            255
-        ).toString();
-
-        const Node = (balanceImage: string) => (
-            <Avatar
-                source={{
-                    uri: balanceImage
-                }}
-                rounded
-                size="medium"
-            />
-        );
-
         return (
             <View
                 style={{
@@ -140,7 +101,10 @@ export default class Settings extends React.Component<SettingsProps, {}> {
                         >
                             {selectedNode && (
                                 <View style={{ padding: 0 }}>
-                                    {Node(`data:image/png;base64,${data}`)}
+                                    <NodeIdenticon
+                                        selectedNode={selectedNode}
+                                        width={50}
+                                    />
                                 </View>
                             )}
                             <Text
@@ -151,7 +115,7 @@ export default class Settings extends React.Component<SettingsProps, {}> {
                                 }}
                             >
                                 {selectedNode
-                                    ? displayName
+                                    ? NodeTitle(selectedNode)
                                     : localeString(
                                           'views.Settings.connectNode'
                                       )}
